Delete all legacy dashboard frontend code.
Migrate some APIs to use the new dashboard API libraries and type definitions

Remove libraries that was only used by old dashboard: react-redux, numeral, use-debounce
