+++ dashboard/client/package.json
@@ -24,16 +24,13 @@
     "dayjs": "^1.9.4",
     "lodash": "^4.17.20",
     "lowlight": "^1.14.0",
-    "numeral": "^2.0.6",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-redux": "^7.2.0",
     "react-router-dom": "^5.1.2",
     "react-scripts": "^5.0.1",
     "react-window": "^1.8.5",
     "typeface-roboto": "0.0.75",
-    "typescript": "^4.8.4",
-    "use-debounce": "^3.4.3"
+    "typescript": "^4.8.4"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.5",
