+++ dashboard/client/src/service/global.ts
@@ -1,0 +1,9 @@
+import { get } from "./requestHandlers";
+
+export type UsageStatsEnabledResponse = {
+  usageStatsEnabled: boolean;
+  usageStatsPromptEnabled: boolean;
+};
+
+export const getUsageStatsEnabled = () =>
+  get<UsageStatsEnabledResponse>("/usage_stats_enabled");
